researcher: make all m_total_magnitude double->uint64_t casts explicit
The earlier commit in this PR widened ResearchAccount::m_total_magnitude
from uint32_t to uint64_t and made the truncating cast explicit at the
RecordRewardBlock site, but Copilot review on PR #2957 caught that the
three other call sites still relied on the implicit floating->unsigned
conversion (which is UB when out-of-range and silently truncates the
fractional magnitude in any case):

  tally.cpp:332  RecordMRCRewardBlock     += pprev->Magnitude()
  tally.cpp:441  Forget                   -= pindex->Magnitude()
  tally.cpp:491  ForgetMRCRewardBlock     -= pprev->Magnitude()

Wrap each in static_cast<uint64_t>(...) so all four +=/-= sites have
matching, explicit conversions. m_total_magnitude / m_accuracy back the
"Lifetime Magnitude Sum" / AverageLifetimeMagnitude reporting only, not
accrual or consensus, so widening + truncating fractional component is
a reporting-fidelity choice, not a correctness one. (NetworkTally's
unrelated uint32_t m_total_magnitude at tally.cpp:164 takes a uint64_t
assignment from SuperblockCpids::TotalMagnitude() -- a narrowing
integer-to-integer conversion, not a double-to-unsigned UB candidate,
and the network-wide total is well within uint32_t range; out of scope.)

+++ src/gridcoin/tally.cpp
@@ -329,7 +329,7 @@ public:
             // used in accrual calculations, only reporting.
             if (mrc_researcher->m_magnitude > 0) {
                 account.m_accuracy++;
-                account.m_total_magnitude += pindex->pprev->Magnitude();
+                account.m_total_magnitude += static_cast<uint64_t>(pindex->pprev->Magnitude());
             }
 
             if (account.m_first_block_ptr == nullptr) {
@@ -438,7 +438,7 @@ public:
 
         if (pindex->Magnitude() > 0) {
             account.m_accuracy--;
-            account.m_total_magnitude -= pindex->Magnitude();
+            account.m_total_magnitude -= static_cast<uint64_t>(pindex->Magnitude());
         }
 
         pindex = FindLastRewardBlock(cpid, pindex);
@@ -488,7 +488,7 @@ public:
 
             if (mrc_researcher->m_magnitude > 0) {
                 account.m_accuracy--;
-                account.m_total_magnitude -= pindex->pprev->Magnitude();
+                account.m_total_magnitude -= static_cast<uint64_t>(pindex->pprev->Magnitude());
             }
 
             const CBlockIndex* last_block_pindex = FindLastRewardBlock(cpid, pindex);
